requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.1
Flask-Login==0.6.3
pymongo==4.6.1
pandas==2.2.0
pyotp
qrcode[pil]
Pillow


# Testing
pytest==8.0.0
pytest-flask==1.3.0

# Optional but recommended for security and formatting
email-validator==2.1.0.post1
python-dotenv==1.0.1
